requirements.txt
pyusb==1.2.1
pystray==0.19.5
pillow==10.4.0
schedule~=1.2.2